Rename base module folder
