+++ src/base/encoder.c
