+++ src/base/encoder.h
