+++ src/base/position_manager.c
