+++ tests/polar_test.cpp
@@ -1,6 +1,6 @@
 #include <CppUTest/TestHarness.h>
 #include <math.h>
-#include "odometry/polar.h"
+#include "base/polar.h"
 
 TEST_GROUP(Polar)
 {
